pyproject.toml
[project]
name = "weblog"
version = "0.1.0"
description = "A blog built with Wagtail"
dependencies = [
    "wagtail>=5.2.3",
    "wagtail-modeladmin>=2.0.0",
    "django>=4.2.0",
    "django-taggit>=5.0.1",
    "markdown>=3.5.2",
    "Pillow>=10.2.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["srcs/weblog"] 